ref: migrated last test to new setup
